Add helpful navigation commands
The previous example missed some moving around steps.
Added .git removal after each clone so we can eliminate step 6.
Don't cd to the themes directory when cloning Sage so we don't need to move back. Stay in the project root.
Reordered the steps so the cloning steps are grouped together, and the `mv` step comes last. If the clone commands are too long, could move all the `rm -rf .git` tasks to the end.