Update README.md to mention ANSIBLE_PATH
By following the installation instructions in this readme the ANSIBLE_PATH in the Vagrantfile defaults to `'.'` and gives the following error on `vagrant up`:
```
There was an error loading a Vagrantfile. The file being loaded
and the error message are shown below. This is usually caused by
a syntax error.

Path: /roots-example-project/Vagrantfile
Message: ./group_vars/development file not found. Please set `ANSIBLE_PATH` in Vagrantfile
```

This PR updates the readme to include instructions to change the ANSIBLE_PATH in the Vagrantfile